whisper-transcription/pyproject.toml
[project]
name = "practice-asobi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.27.0",
    "numpy==1.26.4",
    "pandas>=1.3.0",
    "pydub>=0.25.1",
    "librosa>=0.10.0",
    "soundfile>=0.12.1",
    "openai>=2.14.0",
    "moviepy>=2.2.1",
    "python-dotenv>=1.0.0",
    "langchain-openai>=1.1.7",
    "langchain-core>=1.2.7",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
]
